refactor(e2e): retire the last of the func() bool waits
AnyPodShouldBeRunning and DeploymentAvailable were the last two, and
nothing calls them now. conditions.go is empty with them gone, so it and
its test file go too; the two helpers condition_test.go still needs move
across.

internal/wait is now Conditions and nothing else.
